refactor(Schedule): group schedule components into its own component directory
