+++ components/modals/ClassInfo/ClassInfo.tsx
@@ -3,7 +3,7 @@ import AccessTimeRoundedIcon from "@mui/icons-material/AccessTimeRounded";
 import { simpleTimeStringFromISO, WEEKDAY_NUMBER_TO_NAME } from "../../../utils/timeUtils";
 import LocationOnRoundedIcon from "@mui/icons-material/LocationOnRounded";
 import PersonRoundedIcon from "@mui/icons-material/PersonRounded";
-import ClassPopularityMeter from "../../ClassCard/ClassPopularityMeter";
+import ClassPopularityMeter from "../../schedule/class/ClassPopularityMeter";
 import Image from "next/image";
 import React, { useState } from "react";
 import { SitClass } from "../../../types/sitTypes";
@@ -13,7 +13,7 @@ import { SIT_TIMEZONE } from "../../../config/config";
 import { formatNameArray } from "../../../utils/arrayUtils";
 import { ClassPopularity, SessionStatus, StatusColors, UserNameWithIsSelf } from "../../../types/rezervoTypes";
 import { stringifyClassPopularity } from "../../../lib/popularity";
-import ClassUsersAvatarGroup from "../../ClassUsersAvatarGroup";
+import ClassUsersAvatarGroup from "../../schedule/class/ClassUsersAvatarGroup";
 import LoadingButton from "@mui/lab/LoadingButton";
 import ConfirmationDialog from "../../ConfirmationDialog";
 import { useUserSessions } from "../../../hooks/useUserSessions";
