+++ components/schedule/WeekSchedule.tsx
@@ -1,8 +1,8 @@
 import { Box, Stack } from "@mui/material";
 import DaySchedule from "./DaySchedule";
 import React, { useEffect } from "react";
-import { AllConfigsIndex, ClassPopularityIndex } from "../types/rezervoTypes";
-import { SitClass, SitSchedule } from "../types/sitTypes";
+import { AllConfigsIndex, ClassPopularityIndex } from "../../types/rezervoTypes";
+import { SitClass, SitSchedule } from "../../types/sitTypes";
 import { useRouter } from "next/router";
 
 function WeekSchedule({
