+++ components/schedule/class/ClassCard.tsx
@@ -1,18 +1,18 @@
 import { Avatar, AvatarGroup, Box, Card, CardActions, CardContent, Typography } from "@mui/material";
 import React, { useEffect, useState } from "react";
-import { SitClass } from "../../types/sitTypes";
-import { simpleTimeStringFromISO } from "../../utils/timeUtils";
-import { hexColorHash, hexWithOpacityToRgb } from "../../utils/colorUtils";
-import { EnterLeaveAnimation, OVER_THE_TOP_ANIMATIONS } from "../../types/animationTypes";
-import { randomElementFromArray } from "../../utils/arrayUtils";
+import { SitClass } from "../../../types/sitTypes";
+import { simpleTimeStringFromISO } from "../../../utils/timeUtils";
+import { hexColorHash, hexWithOpacityToRgb } from "../../../utils/colorUtils";
+import { EnterLeaveAnimation, OVER_THE_TOP_ANIMATIONS } from "../../../types/animationTypes";
+import { randomElementFromArray } from "../../../utils/arrayUtils";
 import IconButton from "@mui/material/IconButton";
 import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined";
 import ClassPopularityMeter from "./ClassPopularityMeter";
-import { ClassPopularity, SessionStatus, StatusColors, UserNameWithIsSelf } from "../../types/rezervoTypes";
-import RippleBadge from "../RippleBadge";
-import { isClassInThePast } from "../../lib/iBooking";
-import ClassUserAvatar from "../ClassUserAvatar";
-import { useUserSessions } from "../../hooks/useUserSessions";
+import { ClassPopularity, SessionStatus, StatusColors, UserNameWithIsSelf } from "../../../types/rezervoTypes";
+import RippleBadge from "../../RippleBadge";
+import { isClassInThePast } from "../../../lib/iBooking";
+import ClassUserAvatar from "./ClassUserAvatar";
+import { useUserSessions } from "../../../hooks/useUserSessions";
 
 const ClassCard = ({
     _class,
