+++ components/schedule/class/ClassUsersAvatarGroup.tsx
